tests for BridgeHub <> remote GRANDPA chain
