+++ cumulus/parachains/runtimes/bridge-hubs/bridge-hub-rococo/tests/tests.rs
@@ -26,7 +26,6 @@ use bridge_hub_rococo_runtime::{
 };
 use codec::{Decode, Encode};
 use frame_support::{dispatch::GetDispatchInfo, parameter_types};
-use frame_system::pallet_prelude::HeaderFor;
 use parachains_common::{rococo::fee::WeightToFee, AccountId, AuraId, Balance};
 use sp_keyring::AccountKeyring::Alice;
 use sp_runtime::{
@@ -46,13 +45,16 @@ fn construct_extrinsic(
 	sender: sp_keyring::AccountKeyring,
 	call: RuntimeCall,
 ) -> UncheckedExtrinsic {
+	let account_id = AccountId32::from(sender.public());
 	let extra: SignedExtra = (
 		frame_system::CheckNonZeroSender::<Runtime>::new(),
 		frame_system::CheckSpecVersion::<Runtime>::new(),
 		frame_system::CheckTxVersion::<Runtime>::new(),
 		frame_system::CheckGenesis::<Runtime>::new(),
 		frame_system::CheckEra::<Runtime>::from(Era::immortal()),
-		frame_system::CheckNonce::<Runtime>::from(0),
+		frame_system::CheckNonce::<Runtime>::from(
+			frame_system::Pallet::<Runtime>::account(&account_id).nonce
+		),
 		frame_system::CheckWeight::<Runtime>::new(),
 		pallet_transaction_payment::ChargeTransactionPayment::<Runtime>::from(0),
 		BridgeRejectObsoleteHeadersAndMessages::default(),
@@ -62,7 +64,7 @@ fn construct_extrinsic(
 	let signature = payload.using_encoded(|e| sender.sign(e));
 	UncheckedExtrinsic::new_signed(
 		call,
-		AccountId32::from(sender.public()).into(),
+		account_id.into(),
 		Signature::Sr25519(signature.clone()),
 		extra,
 	)
@@ -70,10 +72,9 @@ fn construct_extrinsic(
 
 fn construct_and_apply_extrinsic(
 	relayer_at_target: sp_keyring::AccountKeyring,
-	batch: pallet_utility::Call<Runtime>,
+	call: RuntimeCall,
 ) -> sp_runtime::DispatchOutcome {
-	let batch_call = RuntimeCall::Utility(batch);
-	let xt = construct_extrinsic(relayer_at_target, batch_call);
+	let xt = construct_extrinsic(relayer_at_target, call);
 	let r = Executive::apply_extrinsic(xt);
 	r.unwrap()
 }
@@ -83,10 +84,6 @@ fn construct_and_estimate_extrinsic_fee(batch: pallet_utility::Call<Runtime>) ->
 	let batch_info = batch_call.get_dispatch_info();
 	let xt = construct_extrinsic(Alice, batch_call);
 	TransactionPayment::compute_fee(xt.encoded_size() as _, &batch_info, 0)
-}
-
-fn executive_init_block(header: &HeaderFor<Runtime>) {
-	Executive::initialize_block(header)
 }
 
 fn collator_session_keys() -> bridge_hub_test_utils::CollatorSessionKeys<Runtime> {
@@ -235,10 +232,9 @@ mod bridge_hub_rococo_tests {
 	#[test]
 	fn relayed_incoming_message_works() {
 		// from Westend
-		bridge_hub_test_utils::test_cases::relayed_incoming_message_works::<
+		bridge_hub_test_utils::test_cases::from_parachain::relayed_incoming_message_works::<
 			Runtime,
 			AllPalletsWithoutSystem,
-			XcmConfig,
 			ParachainSystem,
 			BridgeGrandpaWestendInstance,
 			BridgeParachainWestendInstance,
@@ -248,17 +244,19 @@ mod bridge_hub_rococo_tests {
 			collator_session_keys(),
 			bp_bridge_hub_rococo::BRIDGE_HUB_ROCOCO_PARACHAIN_ID,
 			bp_bridge_hub_westend::BRIDGE_HUB_WESTEND_PARACHAIN_ID,
+			BridgeHubWestendChainId::get(),
 			SIBLING_PARACHAIN_ID,
 			Rococo,
 			XCM_LANE_FOR_ASSET_HUB_ROCOCO_TO_ASSET_HUB_WESTEND,
 			|| (),
+			construct_and_apply_extrinsic,
 		)
 	}
 
 	#[test]
 	pub fn complex_relay_extrinsic_works() {
 		// for Westend
-		bridge_hub_test_utils::test_cases::complex_relay_extrinsic_works::<
+		bridge_hub_test_utils::test_cases::from_parachain::complex_relay_extrinsic_works::<
 			Runtime,
 			AllPalletsWithoutSystem,
 			XcmConfig,
@@ -275,10 +273,8 @@ mod bridge_hub_rococo_tests {
 			BridgeHubWestendChainId::get(),
 			Rococo,
 			XCM_LANE_FOR_ASSET_HUB_ROCOCO_TO_ASSET_HUB_WESTEND,
-			ExistentialDeposit::get(),
-			executive_init_block,
-			construct_and_apply_extrinsic,
 			|| (),
+			construct_and_apply_extrinsic,
 		);
 	}
 
@@ -302,7 +298,7 @@ mod bridge_hub_rococo_tests {
 
 	#[test]
 	pub fn can_calculate_fee_for_complex_message_delivery_transaction() {
-		let estimated = bridge_hub_test_utils::test_cases::can_calculate_fee_for_complex_message_delivery_transaction::<
+		let estimated = bridge_hub_test_utils::test_cases::from_parachain::can_calculate_fee_for_complex_message_delivery_transaction::<
 			Runtime,
 			BridgeGrandpaWestendInstance,
 			BridgeParachainWestendInstance,
@@ -325,7 +321,7 @@ mod bridge_hub_rococo_tests {
 
 	#[test]
 	pub fn can_calculate_fee_for_complex_message_confirmation_transaction() {
-		let estimated = bridge_hub_test_utils::test_cases::can_calculate_fee_for_complex_message_confirmation_transaction::<
+		let estimated = bridge_hub_test_utils::test_cases::from_parachain::can_calculate_fee_for_complex_message_confirmation_transaction::<
 			Runtime,
 			BridgeGrandpaWestendInstance,
 			BridgeParachainWestendInstance,
